design(stock): split /stock into Overview + Products routes
User feedback: dual-tab toggle (ภาพรวม/รายการ) on /stock was hard to
use. Splitting follows the pattern of /stock/transfers, /stock/alerts,
/stock/adjustments, /stock/count, /stock/workflow which are already
separate routes.

Routes
- /stock          → StockOverviewPage (dashboard only)
- /stock/products → StockProductsPage (list + filters + bulk actions)
- /products       → redirect /stock/products (was: /stock?tab=list)

Hook split
- useStockOverview(branchId) — summary, dashboard, warranty, branches
- useStockProducts() — list + URL-backed filters + selection +
  bulk transfer + price modal + export

Filter state moved to URL params (?branchId / ?status / ?category /
?q / ?page) so deep-links from ActionZone work and bookmarks preserve
context across navigation.

Sidebar (4 places: SALES, BRANCH_MANAGER, OWNER + mobile bottomNav):
- 'ภาพรวมคลัง' → /stock
- 'รายการสินค้า' → /stock/products  (NEW)
Mobile shortcut keeps /stock as the entry.

Removed
- StockPage/index.tsx (replaced by 2 page files)
- hooks/useStockData.ts (split)
- hooks/useStockFilters.ts (URL-backed in useStockProducts)

Spec: docs/superpowers/specs/2026-05-08-stock-page-split-design.md
