+++ apps/web/src/config/menu.ts
@@ -104,7 +104,8 @@ const SALES_CONFIG: RoleMenuConfig = {
       label: 'เครื่องมือ',
       icon: Warehouse,
       items: [
-        { label: 'สต็อกสินค้า', path: '/stock', icon: Warehouse },
+        { label: 'ภาพรวมคลัง', path: '/stock', icon: Warehouse },
+        { label: 'รายการสินค้า', path: '/stock/products', icon: ClipboardList },
         { label: 'พิมพ์สติกเกอร์', path: '/stickers', icon: Tag },
         { label: 'ค่าคอมมิชชัน', path: '/commissions', icon: Coins },
         { label: 'CRM Pipeline', path: '/crm', icon: Kanban },
@@ -154,7 +155,8 @@ const BRANCH_MANAGER_CONFIG: RoleMenuConfig = {
       label: 'คลัง & จัดซื้อ',
       icon: Warehouse,
       items: [
-        { label: 'สต็อกสินค้า', path: '/stock', icon: Warehouse },
+        { label: 'ภาพรวมคลัง', path: '/stock', icon: Warehouse },
+        { label: 'รายการสินค้า', path: '/stock/products', icon: ClipboardList },
         { label: 'พิมพ์สติกเกอร์', path: '/stickers', icon: Tag },
         { label: 'สั่งซื้อ (PO)', path: '/purchase-orders', icon: ClipboardList },
         { label: 'ผู้ขาย', path: '/suppliers', icon: Building2 },
@@ -333,7 +335,8 @@ const OWNER_CONFIG: RoleMenuConfig = {
         { label: 'ผู้ขาย', path: '/suppliers', icon: Building2 },
         { label: 'สั่งซื้อ (PO)', path: '/purchase-orders', icon: ClipboardList },
         { label: 'รับซื้อมือสอง', path: '/trade-in', icon: Smartphone },
-        { label: 'สต็อกสินค้า', path: '/stock', icon: Warehouse },
+        { label: 'ภาพรวมคลัง', path: '/stock', icon: Warehouse },
+        { label: 'รายการสินค้า', path: '/stock/products', icon: ClipboardList },
         { label: 'พิมพ์สติกเกอร์', path: '/stickers', icon: Tag },
       ],
     },
